+++ src/components/date-range-picker/date-range-picker.utils.spec.ts
@@ -1,11 +1,14 @@
 import { elementUpdated, expect } from '@open-wc/testing';
+import type { CalendarDay } from '#internals/date/model.js';
 import { formatDisplayDate } from '#internals/i18n/i18n-controller.js';
+import {
+  getCalendarDOM,
+  getDOMDate,
+} from '#internals/testing/calendar.spec.js';
 import { checkDatesEqual } from '#internals/testing/helpers.spec.js';
 import { simulateClick } from '#internals/testing/simulate.spec.js';
 import { equal } from '#internals/utils/objects.js';
 import IgcCalendarComponent from '../calendar/calendar.js';
-import { getCalendarDOM, getDOMDate } from '../calendar/helpers.spec.js';
-import type { CalendarDay } from '../calendar/model.js';
 import IgcDateTimeInputComponent from '../date-time-input/date-time-input.js';
 import type { DateRangeValue } from '../types.js';
 import IgcDateRangeInputComponent from './date-range-input.js';